auto change gpu and cpu
